Correct strlen to return a size_t

+++ src/string/ft_strlen.c
@@ -1,6 +1,8 @@
-int	ft_strlen(const char *s) {
-	int	i = 0;
-	
+#include "libft.h"
+
+size_t	ft_strlen(const char *s) {
+	size_t	i = 0;
+
 	while (s && s[i]) {
 		i++;
 	}
